+++ src/lib/test-utils/propTypes.ts
@@ -1,0 +1,11 @@
+import type { PropTypeInfo } from '../props/propDefinitions.types';
+
+const getPropTypeInfo = (propTypeInfo: PropTypeInfo): PropTypeInfo => ({
+  ...propTypeInfo,
+  source: {
+    ...propTypeInfo.source,
+    name: propTypeInfo.source.name || propTypeInfo.name,
+  },
+});
+
+export default getPropTypeInfo;